requirements.txt
beautifulsoup4
duden
pyyaml
